requirements.txt
amqp==5.1.1
asgiref==3.7.2
bcrypt==4.0.1
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.5
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
dnspython==2.4.2
eventlet==0.33.3
greenlet==3.0.0
idna==3.4
isort==5.12.0
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
